requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
h11==0.14.0
idna==3.10
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
pytest==8.3.5
python-dateutil==2.9.0.post0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlmodel==0.0.24
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
uvicorn==0.34.0
